Add PWA Registration, Optional CORS options, Docker file updates, Directory restructures
Update readme.md, gitignore, and .env.example

update save error todo toast timeout

initialize theme on load

update .env.example

clean up dockerfile

update deps
