+++ .env.example
@@ -6,3 +6,9 @@ DUMBDO_PIN=1234
 
 # Server Port (default: 3000)
 PORT=3000 
+
+DUMBDO_SITE_TITLE=DumbDo
+
+# (Optional) Restrict origins - ex: https://subdomain.domain.tld,https://auth.proxy.tld,http://internalip:port' (default is '*')
+# - ALLOWED_ORIGINS=http://localhost:3000
+# NODE_ENV=development # default production (development allows all origins)
