+++ docker-compose.yml
@@ -1,0 +1,21 @@
+services:
+  dumbdo:
+    image: dumbwareio/dumbdo:latest
+    container_name: dumbdo
+    restart: unless-stopped
+    ports:
+      - ${DUMBDO_PORT:-3000}:3000
+    volumes:
+      - ${DUMBDO_DATA_PATH:-./data}:/app/data
+    environment:
+      - DUMBDO_PIN=${DUMBDO_PIN-}
+      - DUMBDO_SITE_TITLE=DumbDo
+      # (Optional) Restrict origins - ex: https://subdomain.domain.tld,https://auth.proxy.tld,http://internalip:port' (default is '*')
+      # - ALLOWED_ORIGINS=http://localhost:3000
+      # - NODE_ENV=development # default production (development allows all origins)
+    #healthcheck:
+    #  test: wget --spider -q  http://127.0.0.1:3000
+    #  start_period: 20s
+    #  interval: 20s
+    #  timeout: 5s
+    #  retries: 3
