+++ public/assets/styles.css
@@ -8,6 +8,8 @@
     --border: #ccc;
     --shadow: 0 2px 4px rgba(0,0,0,0.1);
     --transition: 0.2s ease;
+    --success-status-bg: rgba(37, 99, 235, 0.6);
+    --danger-status-bg:rgba(220, 38, 38, 0.6);
 }
 
 [data-theme="dark"] {
@@ -16,6 +18,8 @@
     --text: white;
     --border: #404040;
     --shadow: 0 2px 4px rgba(0,0,0,0.2);
+    --success-status-bg: rgba(96, 165, 250, 0.5);
+    --danger-status-bg:rgba(220, 38, 38, 0.5);
 }
 
 * {
@@ -274,7 +278,7 @@ input:focus {
     background: rgba(255,68,68,0.1);
 }
 
-.toast {
+/* .toast-container {
     position: fixed;
     bottom: 2rem;
     left: 50%;
@@ -285,11 +289,46 @@ input:focus {
     box-shadow: var(--shadow);
     opacity: 0;
     transition: opacity var(--transition);
+    z-index: 2000;
+} */
+
+.toast-container {
+    position: fixed;
+    bottom: 1rem;
+    left: 50%;
+    transform: translateX(-50%);
+    padding: 0.5rem 1rem;
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+    z-index: 2000;
+}
+
+.toast {
+    color: #ffffff;
+    padding: 0.5rem 1rem;
+    border-radius: 20px;
+    opacity: 0;
+    transition: opacity 0.3s ease-in-out;
+    max-width: 300px;
+    box-sizing: border-box;
+    word-wrap: break-word;
+    font-size: 0.875rem;
+    cursor: pointer;
 }
 
 .toast.show {
     opacity: 1;
 }
+
+.toast.success {
+    background-color: var(--success-status-bg);
+}
+
+.toast.error {
+    background-color: var(--danger-status-bg);
+}
+
 
 .todo-divider {
     width: 100%;
