+++ public/index.html
@@ -3,8 +3,9 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>DumbDo - Stupidly Simple Todo List</title>
+    <title id="page-title">DumbDo - Stupidly Simple Todo List</title>
     <link rel="icon" type="image/svg+xml" href="favicon.svg">
+    <link rel="manifest" href="manifest.json">
     <script>
         // Prevent theme flicker by setting theme immediately
         (function() {
@@ -17,7 +18,7 @@
 <body>
     <div class="app">
         <header>
-            <h1>DumbDo</h1>
+            <h1 id="header-title">DumbDo</h1>
             <div class="list-controls">
                 <div class="list-selector-container">
                     <select id="listSelector" aria-label="Select todo list">
@@ -80,9 +81,9 @@
             </ul>
         </main>
 
-        <div id="toast" class="toast" role="alert" aria-hidden="true"></div>
+        <div id="toast-container" class="toast-container"></div>
     </div>
 
-    <script src="app.js"></script>
+    <script src="app.js" type="module"></script>
 </body>
 </html>
