+++ public/login.html
@@ -20,7 +20,7 @@
             <h1>DumbDo</h1>
             <div class="login-box">
                 <h2>Enter PIN</h2>
-                <p class="pin-description">Please enter your PIN to access your todos.</p>
+                <p id="pin-description" class="pin-description">Please enter your PIN to access your todos.</p>
                 <form id="loginForm">
                     <div class="pin-input-container">
                         <!-- PIN inputs will be generated dynamically -->
